# swagger.yml
swagger: '2.0'
info:
  title: Action Exporter API
  description: >
    The Action Exporter API provides support endpoints to store FreeMarker templates, template mappings and to enable
    exports to be initiated for ActionRequests previously stored.

    Note that all of the endpoints this service provides are authenticated using HTTP basic authentication.
  version: 9.32.0
schemes:
  - http
basePath: /
produces:
  - application/json
...
paths:
...
  /templates/{templateName}:
    get:
        summary:  Gets the template for the specified templateName.
        description: Returns details of the template for the supplied templateName.
        parameters:
          - name: Accept
            in: header
            description: Response type, application/json.
            required: false
            type: string
          - name: templateName
            description: Unique template name
            in: path
            required: true
            type: string
        tags:
          - Template
        responses:
          200:
            description: Template details for the supplied templateName.
            schema:
              $ref: '#/definitions/Template'
          404:
            description: templateName not found
            schema:
              $ref: '#/definitions/Error'
          default:
            description: Unexpected error
            schema:
              $ref: '#/definitions/Error'
    post:
      summary: Stores a FreeMarker template
      description: Stores a FreeMarker template
      consumes:
        - multipart/form-data
      parameters:
        - name: Accept
          in: header
          description: Response type, application/json
          required: false
          type: string
        - name: templateName
          in: path
          type: string
          description: Unique template name
          required: true
        - name: file
          in: formData
          type: file
          description: Template file to store
          required: true
      tags:
        - Template
      responses:
        201:
          description: Template file stored
          schema:
            $ref: '#/definitions/Template'
        400:
          description: Invalid input
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
...
definitions:
...
  Template:
    type: object
    required:
      - name
      - content
      - dateModified
    properties:
      name:
        type: string
        description: Unique name for the Template.
      content:
        type: string
        description: Content of the Template.
      dateModified:
        type: string
        format: date-time
        description: Creation date-time.
...
  Error:
    type: object
    properties:
      code:
        description: Error code
        type: string
      timestamp:
        description:  Unique error timestamp that can be searched for in logs
        type: integer
      message:
        description: Error message
        type: string
...
securityDefinitions:
  UserSecurity:
    type: basic
security:
  - UserSecurity: []